// frontend/src/groups/serviceaccount/__generated__/ServiceAccountListFragment_serviceAccounts.graphql.ts
export type ServiceAccountListFragment_serviceAccounts$data = {
  readonly id: string;
  readonly serviceAccounts: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly groupPath: string;
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"ServiceAccountListItemFragment_serviceAccount">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "ServiceAccountListFragment_serviceAccounts";
};
export type ServiceAccountListFragment_serviceAccounts$key = {
  readonly " $data"?: ServiceAccountListFragment_serviceAccounts$data;
  readonly " $fragmentSpreads": FragmentRefs<"ServiceAccountListFragment_serviceAccounts">;
};